+++ tests/test_get_collisions.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+
+import pytest
+import subprocess as sub
+from bin import get_collisions
+
+def test_noargs():
+    sub.call(get_collisions.__file__, shell=True)
+    
+def test_oor():
+    sub.call([get_collisions.__file__, "-o"], shell=True)
+
+if __name__ == "__main__":
+    pytest.main()
+